test longer distance

+++ src/library/ultrasonic.h
@@ -14,5 +14,5 @@ void setReceive_listener(listener event);
 long int set_cycle(int ticks, int cycle);
 
 static int FLAG=0;
-static long int OUT_NUM=0;
+static u32 OUT_NUM=0;
 static uint32_t ULTRA_EMIT=0;//used to store hardware emit signal time
